Add potential logic for storing holders.
First experimentation with making the calculator
work by storing the displayValues in holder
variables and adding (so far) the numbers together.

So far only works with adding 2 digits, does not
support addition for 2-digit+ numbers nor does it
update the display properly during the calculation.

It does add though, which is a bonus!

Will continue testing and adding more logic
tomorrow.